apps/backend/requirements.txt
vrchatapi>=1.18.0
pyotp>=2.9.0
psycopg2-binary>=2.9.9
schedule>=1.2.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
